src/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "demo-mcp-server"
description = "Demo MCP Server with OAuth 2.1 authentication and On-Behalf-Of flow for Microsoft Graph API."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Demo MCP Server Contributors" }
]
dynamic = ["version"]

# Core runtime dependencies (pinned for reproducibility)
dependencies = [
  "fastmcp==2.13.0.2",
  "uvicorn[standard]==0.36.0",
  "python-dotenv==1.1.1",
  "azure-identity==1.19.0",
  "pydantic==2.11.7",
  "pydantic-settings==2.6.1",
  "python-multipart==0.0.18",
  "httpx==0.28.1",
  "PyJWT==2.9.0",
  "aiohttp==3.10.11",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "pytest-asyncio==0.24.0",
  "cryptography>=41.0.0",
  "PyJWT==2.9.0",
]

[project.urls]
Homepage = "https://github.com/huangyingting/mcp-workshop"
Repository = "https://github.com/your-org/demo-mcp-server.git"

# Version is sourced from the package __init__.py
[tool.hatch.version]
path = "__init__.py"

# Hatch build configuration
[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["**/*", "README.md"]

[project.scripts]
demo-mcp-server = "server:main"
